+++ test/Kentico.Kontent.ModelGenerator.Tests/DeliveryCodeGeneratorTests.cs
@@ -38,36 +38,6 @@ namespace Kentico.Kontent.ModelGenerator.Tests
             Assert.Throws<InvalidOperationException>(() => new DeliveryCodeGenerator(mockOptions.Object, outputProvider.Object, deliveryClient.Object));
         }
 
-        [Fact]
-        public void CreateCodeGeneratorOptions_NoOutputSetInJsonNorInParameters_OutputDirHasDefaultValue()
-        {
-            var mockOptions = new Mock<IOptions<CodeGeneratorOptions>>();
-            var options = new CodeGeneratorOptions
-            {
-                OutputDir = ""
-            };
-            mockOptions.Setup(x => x.Value).Returns(options);
-
-            var outputProvider = new FileSystemOutputProvider(mockOptions.Object);
-            Assert.Empty(options.OutputDir);
-            Assert.NotEmpty(outputProvider.OutputDir);
-        }
-
-        [Fact]
-        public void CreateCodeGeneratorOptions_OutputSetInParameters_OutputDirHasCustomValue()
-        {
-            var expectedOutputDir = Environment.CurrentDirectory;
-            var mockOptions = new Mock<IOptions<CodeGeneratorOptions>>();
-            var options = new CodeGeneratorOptions
-            {
-                OutputDir = ""
-            };
-            mockOptions.Setup(x => x.Value).Returns(options);
-
-            var outputProvider = new FileSystemOutputProvider(mockOptions.Object);
-            Assert.Equal(expectedOutputDir.TrimEnd(Path.DirectorySeparatorChar), outputProvider.OutputDir.TrimEnd(Path.DirectorySeparatorChar));
-        }
-
         [Theory]
         [InlineData(true)]
         [InlineData(false)]
